+++ lib/easypost/payload.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'json'
+
+# Webhook Event Payloads are triggered by changes in objects you've created via the API.
+class EasyPost::Payload < EasyPost::Resource
+end